+++ PortfolioViewer/PortfolioViewer.ApiService/appsettings.json
@@ -6,12 +6,12 @@
     }
   },
   "AllowedHosts": "*",
-  "Kestrel": {
-    "Certificates": {
-      "Default": {
-        "Path": "/https/aspnetapp.pfx",
-        "Password": "YourPasswordHere"
-      }
-    }
-  }
+  //"Kestrel": {
+  //  "Certificates": {
+  //    "Default": {
+  //      "Path": "/https/aspnetapp.pfx",
+  //      "Password": "YourPasswordHere"
+  //    }
+  //  }
+  //}
 }
