+++ PortfolioViewer/PortfolioViewer.WASM/PortfolioViewer.WASM.csproj
@@ -33,6 +33,7 @@
 	<ItemGroup>
 		<ProjectReference Include="..\..\Database\Database.csproj" />
 		<ProjectReference Include="..\..\Model\Model.csproj" />
+		<ProjectReference Include="..\PortfolioViewer.Common\PortfolioViewer.Common.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
